Move utils to directory above day-of and pre-check-in to allow both apps to share util code. Add phase one mock script
