+++ src/applications/check-in/day-of/components/BackButton.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import recordEvent from 'platform/monitoring/record-event';
 
-import { createAnalyticsSlug } from '../utils/analytics';
+import { createAnalyticsSlug } from '../../utils/analytics';
 
 const BackButton = props => {
   const { router } = props;
