+++ src/applications/check-in/day-of/containers/withLoadedData.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect, batch, useSelector } from 'react-redux';
 import { compose } from 'redux';
 import { goToNextPage, URLS } from '../utils/navigation';
-import { getCurrentToken } from '../utils/session';
+import { getCurrentToken } from '../../utils/session';
 import { api } from '../api';
 import {
   receivedDemographicsData,
