+++ src/applications/check-in/day-of/pages/CheckIn/DisplayMultipleAppointments.jsx
@@ -7,7 +7,7 @@ import BackButton from '../../components/BackButton';
 import BackToHome from '../../components/BackToHome';
 import Footer from '../../components/Footer';
 import recordEvent from 'platform/monitoring/record-event';
-import { createAnalyticsSlug } from '../../utils/analytics';
+import { createAnalyticsSlug } from '../../../utils/analytics';
 import { sortAppointmentsByStartTime } from '../../utils/appointment';
 import { focusElement } from 'platform/utilities/ui';
