+++ src/applications/check-in/utils/analytics/analytics.test.unit.spec.js
