+++ src/applications/check-in/utils/analytics/index.js
