+++ src/applications/check-in/utils/formatters/index.js
