+++ src/applications/check-in/utils/session/index.js
