+++ src/applications/check-in/utils/session/session.test.unit.spec.js
