+++ src/applications/check-in/utils/token-format-validator/index.js
