+++ src/applications/check-in/utils/token-format-validator/token.format.validator.test.unit.spec.js
